Fixed student viewset
